+++ lib/src/realm_class.dart
@@ -124,6 +124,7 @@ export 'session.dart'
         SyncSessionErrorCode;
 export 'subscription.dart' show Subscription, SubscriptionSet, SubscriptionSetState, MutableSubscriptionSet;
 export 'user.dart' show User, UserState, ApiKeyClient, UserIdentity, ApiKey, FunctionsClient;
+export 'mongodb_collection.dart' show MongoDBClient, MongoDBDatabase, MongoDBCollection;
 export 'native/realm_core.dart' show Decimal128;
 
 /// A [Realm] instance represents a `Realm` database.